Improve Counters
 * Add long press event.
 * Display total fees below the counters.
 * Add step field to CounterBloc.
 * Use generic type for CounterBloc's State.
 * Convert fees to double instead of int.
 * Allow incrementation only if the user's balance is enough.

+++ lib/blocs/qr_bloc/qr_event.dart
@@ -5,7 +5,7 @@ class QrEvent extends Equatable {
   final String id;
   final String secret;
   final int quantity;
-  final int fees;
+  final double fees;
   String data;
 
   QrEvent({
